+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+### Fixed
+- Fix scroll directions of scroll bar arrow in instrument sequence editor
 
 ## v0.1.0 (2018-11-25)
 ### Added